src/backend/pyproject.toml
[project]
name = "matrix-backend"
version = "0.1.0"
description = "Matrix — Competency Matrix Platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "redis>=5.0",
    "celery>=5.4",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "email-validator>=2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["B008", "UP042", "UP046"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["../../tests/backend"]
pythonpath = ["."]

[tool.coverage.run]
concurrency = ["greenlet", "thread"]
omit = [
    "app/tasks/backup.py",
    "app/telegram_bot.py",
    "app/cli.py",
    "seed.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
